Create post project carousel and create initial Eagle Project layout
This commit puts forth the basic way I want to display the eagle scout project.
I will be using flex to do the following:

On Desktop I will have the headers "Before Pictures" and "After Pictures"
be placed on top of their corresponding slideshows in a 2x2 grid. When the site
is viewed on a smaller device, the layout changes into a stacked formation, where
"Before Pictures" will be on top of its slidehow, which will be on top of
"After Pictures", which will be on top of  its slideshow.